Rearchitect the entire thing
Ensure now is just a tiny executable that runs a bunch of plugins. As
long as your path includes an `ensure-foo`, you can `ensure foo`.

This makes it extensible by users into adding anything. Wanna make it
work with debian? Just make sure you have an `ensure-package` and
`ensure-services` in your path that use apt and upstart instead of
homebrew and launchctl.